SBERDOMA-63 packages/@core.codegen/templates change to domains based folders
