--- packages/@core.codegen/templates/app00/access/README.md
@@ -1,6 +1,0 @@
-This directory contains all access checks for all models and services used on the server side. 
-
-The folder structure is similar to the `/schema` folder.
-You can read more detail about domain logic in the `/schema/README.md` file.
-
-Each domain have own file.
