--- packages/@core.codegen/templates/app00/constants/README.md
@@ -1,7 +1,0 @@
-The directory contains various constants used in the client side and in the server side.
-For example, it can be validation error codes, regular expressions, and so on.
-
-IF it's the domain logic constants please put it inside a <domain-name> file.
-You can read more detail about domain logic in the `/schema/README.md` file.
-
-Each domain have own file.
