+++ packages/@core.codegen/templates/app00/domains/common/access/README.md
@@ -1,0 +1,3 @@
+This directory contains all access checks for all models and services used on the server side. 
+
+You can read more detail about domain logic in the `/domains/README.md` file.
