+++ packages/@core.codegen/templates/app00/domains/common/constants/README.md
@@ -1,0 +1,6 @@
+The directory contains various constants used in the client side and in the server side.
+For example, it can be validation error codes, regular expressions, and so on.
+
+IF it's the domain logic constants please put it inside a `domains/<domain-name>/constants` file.
+
+You can read more detail about domain logic in the `/domains/README.md` file.
