+++ packages/@core.codegen/templates/app00/domains/common/constants/errors.js
