+++ packages/@core.codegen/templates/app00/domains/common/constants/regexps.js
