+++ packages/@core.codegen/templates/app00/domains/common/gql/README.md
@@ -1,0 +1,3 @@
+This directory contains all GraphQL queries used on the server and client side. 
+
+You can read more detail about domain logic in the `/domains/README.md` file.
