+++ packages/@core.codegen/templates/app00/domains/common/schema/README.md
@@ -1,0 +1,3 @@
+Each **model** have **own file** with the same name as the model. 
+
+All model files are located inside the domain directory.
