+++ packages/@core.codegen/templates/app00/domains/common/utils/README.md
@@ -1,0 +1,9 @@
+All reusable function which is not related to domain logic.
+
+If it's client side logic utils function please put it inside a related 
+`/<domain>/components` component.
+
+We also have specific domain utilities and helpers 
+for server side domain logic (`serverSchema`),
+client side domain data-fetching logic (`clientSchema`) and 
+server side tests logic utilities (`testSchema`)
