+++ packages/@core.codegen/templates/app00/domains/common/utils/serverSchema/README.md
@@ -1,0 +1,3 @@
+This directory contains all server side domain logic utilities.
+
+You can read more detail about domain logic in the `/domains/README.md` file.
