+++ packages/@core.codegen/templates/app00/domains/package.json
@@ -1,0 +1,4 @@
+{
+  "private": true,
+  "version": "1.0.0"
+}
