--- packages/@core.codegen/templates/app00/gql/README.md
@@ -1,6 +1,0 @@
-This directory contains all GraphQL queries used on the server and client side. 
-
-The folder structure is similar to the `/schema` folder.
-You can read more detail about domain logic in the `/schema/README.md` file.
-
-Each domain have own file.
