--- packages/@core.codegen/templates/app00/types/README.md
@@ -1,6 +1,0 @@
-This directory contains all type script types. 
-
-The folder structure is similar to the `/schema` folder.
-You can read more detail about domain logic in the `/schema/README.md` file.
-
-Each domain have own folder.
