--- packages/@core.codegen/templates/app00/types/_common.d.ts
@@ -1,12 +1,0 @@
-export interface IConnectData {
-    connect: IRelationshipField
-}
-
-export interface IRelationshipField {
-    id: String | Number
-}
-
-export interface ISenderField {
-    dv: Number
-    fingerprint: String
-}
