--- packages/@core.codegen/templates/app00/utils/serverSchema/README.md
@@ -1,6 +1,0 @@
-This directory contains all server side domain logic utilities.
-
-The folder structure is similar to the `/schema` folder.
-You can read more detail about domain logic in the `/schema/README.md` file.
-
-Each domain have own folder.
