+++ packages/@core.codegen/templates/schema00/domains/common/schema/fields.js.default
