+++ packages/@core.codegen/templates/schema00/domains/common/utils/codegeneration/generate.hooks.ts
