+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/access/{{name}}.js
