+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/constants.js.default
