+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/gql.js.default
@@ -1,0 +1,15 @@
+/**
+ * Generated by `{{ command }}`
+ * In most cases you should not change it by hands
+ * Please, don't remove `AUTOGENERATE MARKER`s
+ */
+
+const { generateGqlQueries } = require('@domains/common/utils/codegeneration/generate.gql')
+
+const COMMON_FIELDS = 'id dv v deletedAt newId createdBy { id name } updatedBy { id name } createdAt updatedAt'
+
+/* AUTOGENERATE MARKER <CONST> */
+
+module.exports = {
+/* AUTOGENERATE MARKER <EXPORTS> */
+}
