+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/gql.js.patch
