+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/schema/index.js.default
