+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/schema/{{name}}.test.js
@@ -1,0 +1,11 @@
+/**
+ * Generated by `{{ command }}`
+ */
+
+const { makeLoggedInClient, makeLoggedInAdminClient, makeClient } = require('@core/keystone/test.utils')
+
+describe('check access for {{ name }}', () => {
+    test('anonymous', async () => {
+        const client = makeClient()
+    })
+})
