+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/utils/clientSchema/index.ts.patch
