+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/utils/serverSchema/index.js.default
@@ -1,0 +1,15 @@
+/**
+ * Generated by `{{ command }}`
+ * In most cases you should not change it by hands
+ * Please, don't remove `AUTOGENERATE MARKER`s
+ */
+
+const { generateServerUtils } = require('@domains/common/utils/codegeneration/generate.server.utils')
+
+/* AUTOGENERATE MARKER <IMPORT> */
+
+/* AUTOGENERATE MARKER <CONST> */
+
+module.exports = {
+/* AUTOGENERATE MARKER <EXPORTS> */
+}
