+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/utils/serverSchema/index.js.patch
@@ -1,0 +1,11 @@
+/* AUTOGENERATE MARKER <IMPORT> */
+const { {{ name }}: {{ name }}GQL } = require('@domains/{{ domain }}/gql')
+/**/
+
+/* AUTOGENERATE MARKER <CONST> */
+const {{ name }} = generateServerUtils({{ name }}GQL)
+/**/
+
+/* AUTOGENERATE MARKER <EXPORTS> */
+    {{ name }},
+/**/
