+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/utils/testSchema/index.js.default
@@ -1,0 +1,17 @@
+/**
+ * Generated by `{{ command }}`
+ * In most cases you should not change it by hands
+ * Please, don't remove `AUTOGENERATE MARKER`s
+ */
+
+const { generateGQLTestUtils } = require('@domains/common/utils/codegeneration/generate.test.utils')
+
+/* AUTOGENERATE MARKER <IMPORT> */
+
+/* AUTOGENERATE MARKER <CONST> */
+
+/* AUTOGENERATE MARKER <FACTORY> */
+
+module.exports = {
+/* AUTOGENERATE MARKER <EXPORTS> */
+}
