+++ packages/@core.codegen/templates/schema00/domains/{{domain}}/utils/testSchema/index.js.patch
@@ -1,0 +1,34 @@
+/* AUTOGENERATE MARKER <IMPORT> */
+const { {{ name }}: {{ name }}GQL } = require('@domains/{{ domain }}/gql')
+/**/
+
+/* AUTOGENERATE MARKER <CONST> */
+const {{ name }} = generateGQLTestUtils({{ name }}GQL)
+/**/
+
+/* AUTOGENERATE MARKER <FACTORY> */
+async function createTest{{ name }} (client, {% for field in signature %}{% if field[1] === 'Relationship' and not field[0].endsWith('?') %}{{ field[0].replace('?', '') }}, {% endif %}{% endfor %}extraAttrs = {}) {
+    if (!client) throw new Error('no client')
+    const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
+
+    // TODO(codegen): write createTest{{ name }} logic for generate fields
+
+    const attrs = {
+        dv: 1,
+        sender,
+{%- for field in signature %}
+{%- if field[1] === 'Relationship' and not field[0].endsWith('?') %}
+        {{ field[0].replace('?', '') }}: { connect: { id: {{ field[0].replace('?', '') }}.id } },
+{%- endif %}
+{%- endfor %}
+        ...extraAttrs,
+    }
+    const obj = await {{ name }}.create(client, attrs)
+    return [obj, attrs]
+}
+
+/**/
+
+/* AUTOGENERATE MARKER <EXPORTS> */
+    {{ name }}, createTest{{ name }},
+/**/
