--- packages/@core.codegen/templates/schema00/gql/{{domain}}.js.default
@@ -1,15 +1,0 @@
-/**
- * Generated by `{{ command }}`
- * In most cases you should not change it by hands
- * Please, don't remove `AUTOGENERATE MARKER`s
- */
-
-const { generateGqlQueries } = require('../utils/codegeneration/generate.gql')
-
-const COMMON_FIELDS = 'id dv v deletedAt newId createdBy { id name } updatedBy { id name } createdAt updatedAt'
-
-/* AUTOGENERATE MARKER <CONST> */
-
-module.exports = {
-/* AUTOGENERATE MARKER <EXPORTS> */
-}
