--- packages/@core.codegen/templates/schema00/schema/{{domain}}/{{name}}.test.js
@@ -1,55 +1,0 @@
-/**
- * Generated by `{{ command }}`
- */
-
-const { makeLoggedInClient, makeLoggedInAdminClient, makeClient } = require('@core/keystone/test.utils')
-
-describe('check access for {{ name }}', () => {
-    test('anonymous', async () => {
-        const client = makeClient()
-    })
-})
-
-const {{ name }} = new GQLListSchema('{{ name }}', {
-    // TODO(codegen): write doc for the {{ name }} domain model!
-    schemaDoc: 'TODO DOC!',
-    fields: {
-        dv: DV_FIELD,
-        sender: SENDER_FIELD,
-{% for field in signature %}
-        {{ field[0] | replace("?", "") }}: {
-            // TODO(codegen): write doc for {{ name }}.{{ field[0] | replace("?", "") }} field!
-            schemaDoc: 'TODO DOC!',
-            type: {{ field[1] }},
-{%- if field[1] === 'Relationship' %}
-            rel: '{{ field[2] }}',
-{%- elif field[1] === 'Select' %}
-            options: '{{ field[2] }}',
-{%- endif %}
-{%- if field[0].endsWith('?') %}
-{%- else %}
-            isRequired: true,
-{%- endif %}
-{%- if field[1] === 'Relationship' %}
-{%- if field[0].endsWith('?') %}
-{%- else %}
-            knexOptions: { isNotNullable: true }, // Required relationship only!
-{%- endif %}
-            kmigratorOptions: { null: {{ field[0].endsWith('?') }}, on_delete: 'models.{{ field[3] }}' },
-{%- endif %}
-        },
-{% endfor %}
-    },
-    plugins: [uuided(), versioned(), tracked(), softDeleted(), historical()],
-    access: {
-        read: access.canRead{{ pluralize.plural(name) }},
-        create: access.canManage{{ pluralize.plural(name) }},
-        update: access.canManage{{ pluralize.plural(name) }},
-        delete: false,
-        auth: true,
-    },
-})
-
-module.exports = {
-    {{ name }},
-}
