--- packages/@core.codegen/templates/schema00/utils/clientSchema/{{domain}}/{{name}}.ts
@@ -1,53 +1,0 @@
-/**
- * Generated by `{{ command }}`
- */
-
-import { pick, get } from "lodash"
-
-import { getClientSideSenderInfo } from '../../userid.utils'
-import { generateReactHooks } from '../../codegeneration/generate.hooks'
-
-import { {{ name }} } from '../../../gql/{{ domain }}'
-import { I{{ name }}GQLItem, I{{ name }}GQLInput, I{{ name }}UIState, I{{ name }}UIFormState } from "../../../types/{{ domain }}/{{ name }}"
-
-const RELATIONS = [{% for field in signature %}{% if field[1] === 'Relationship' %}'{{ field[0].replace('?', '') }}'{% if not loop.last %}, {% endif %}{% endif %}{% endfor %}]
-
-function convertToUIState (item: I{{ name }}GQLItem): I{{ name }}UIState {
-    if (item.dv !== 1) throw new Error('unsupported item.dv')
-    return pick(item, ['id', 'deletedAt'{% for field in signature %}, '{{ field[0].replace('?', '') }}'{% endfor %}])
-}
-
-function convertToUIFormState (state: I{{ name }}UIState): I{{ name }}UIFormState {
-    const result = {}
-    for (const attr in Object.keys(state)) {
-        result[attr] = (RELATIONS.includes(attr)) ? state[attr].id : state[attr]
-    }
-    return result
-}
-
-function convertToGQLInput (state: I{{ name }}UIFormState | I{{ name }}UIState): I{{ name }}GQLInput {
-    const sender = getClientSideSenderInfo()
-    const result = { dv: 1, sender }
-    for (const attr in Object.keys(state)) {
-        const attrId = get(state[attr], 'id')
-        result[attr] = (RELATIONS.includes(attr)) ? { connect: { id: (attrId || state[attr]) } } : state[attr]
-    }
-    return result
-}
-
-const {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useDelete,
-} = generateReactHooks({{ name }}, { convertToGQLInput, convertToUIState })
-
-export {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useDelete,
-    convertToUIFormState,
-}
