--- packages/@core.codegen/templates/schema00/utils/serverSchema/{{domain}}.js.default
@@ -1,15 +1,0 @@
-/**
- * Generated by `{{ command }}`
- * In most cases you should not change it by hands
- * Please, don't remove `AUTOGENERATE MARKER`s
- */
-
-const { generateServerUtils } = require('../codegeneration/generate.server.utils')
-
-/* AUTOGENERATE MARKER <IMPORT> */
-
-/* AUTOGENERATE MARKER <CONST> */
-
-module.exports = {
-/* AUTOGENERATE MARKER <EXPORTS> */
-}
