--- packages/@core.codegen/templates/schema00/utils/serverSchema/{{domain}}.js.patch
@@ -1,11 +1,0 @@
-/* AUTOGENERATE MARKER <IMPORT> */
-const { {{ name }}: {{ name }}GQL } = require('../../gql/{{ domain }}')
-/**/
-
-/* AUTOGENERATE MARKER <CONST> */
-const {{ name }} = generateServerUtils({{ name }}GQL)
-/**/
-
-/* AUTOGENERATE MARKER <EXPORTS> */
-    {{ name }},
-/**/
