--- packages/@core.codegen/templates/schema00/utils/testSchema/{{domain}}.js.default
@@ -1,17 +1,0 @@
-/**
- * Generated by `{{ command }}`
- * In most cases you should not change it by hands
- * Please, don't remove `AUTOGENERATE MARKER`s
- */
-
-const { generateGQLTestUtils } = require('../codegeneration/generate.test.utils')
-
-/* AUTOGENERATE MARKER <IMPORT> */
-
-/* AUTOGENERATE MARKER <CONST> */
-
-/* AUTOGENERATE MARKER <FACTORY> */
-
-module.exports = {
-/* AUTOGENERATE MARKER <EXPORTS> */
-}
